판매자 판매물품 조회화면 api token사용자의 판매물품조회

+++ src/main/java/com/usit/domain/Product.java
@@ -86,7 +86,7 @@ public class Product implements Serializable {
 	private String tags;
 
 	@Column(name="sell_member_id")
-	private Long sellMemberId;
+	private Integer sellMemberId;
 
 	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
 	@Column(name="reg_date")
